+++ apps/fix_engine/message/test/MessagePool_test.cpp
@@ -57,7 +57,7 @@ public:
         ++_count;
     }
 
-    int _count;
+    size_t _count;
 };
